requirements.txt
accelerate==1.6.0
ackermann-msgs==2.0.2
action-msgs==2.0.2
action-tutorials-interfaces==0.33.4
action-tutorials-py==0.33.4
actionlib-msgs==5.3.5
actuator-msgs==0.0.1
ament-clang-format==0.17.1
ament-clang-tidy==0.17.1
ament-cmake-test==2.5.2
ament-copyright==0.17.1
ament-cppcheck==0.17.1
ament-cpplint==0.17.1
ament-flake8==0.17.1
ament-index-python==1.8.1
ament-lint==0.17.1
ament-lint-cmake==0.17.1
ament-package==0.16.3
ament-pep257==0.17.1
ament-uncrustify==0.17.1
ament-xmllint==0.17.1
angles==1.16.0
apriltag-msgs==2.0.1
bond==4.1.0
builtin-interfaces==2.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
composition-interfaces==2.0.2
contourpy==1.3.2
control-msgs==5.2.0
controller-manager==4.15.0
controller-manager-msgs==4.15.0
cv-bridge==4.1.0
cycler==0.12.1
demo-nodes-py==0.33.4
diagnostic-msgs==5.3.5
diagnostic-updater==4.2.1
diffusers==0.33.1
domain-coordinator==0.12.0
dwb-msgs==1.3.3
example-interfaces==0.12.0
examples-rclpy-executors==0.19.4
examples-rclpy-minimal-action-client==0.19.4
examples-rclpy-minimal-action-server==0.19.4
examples-rclpy-minimal-client==0.19.4
examples-rclpy-minimal-publisher==0.19.4
examples-rclpy-minimal-service==0.19.4
examples-rclpy-minimal-subscriber==0.19.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
generate-parameter-library-py==0.3.8
geographic-msgs==1.0.6
geometry-msgs==5.3.5
gps-msgs==2.0.4
graph-msgs==0.2.0
huggingface-hub==0.30.2
idna==3.10
image-geometry==4.1.0
importlib_metadata==8.6.1
interactive-markers==2.5.4
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.8
laser-geometry==2.7.0
launch==3.4.2
launch-param-builder==0.1.1
launch-ros==0.26.5
launch-testing==3.4.2
launch-testing-ros==0.26.5
launch-xml==3.4.2
launch-yaml==3.4.2
lifecycle-msgs==2.0.2
logging-demo==0.33.4
map-msgs==2.4.1
MarkupSafe==3.0.2
matplotlib==3.10.1
message-filters==4.11.2
moveit-configs-utils==2.11.0
moveit-msgs==2.5.0
mpmath==1.3.0
nav-2d-msgs==1.3.3
nav-msgs==5.3.5
nav2-common==1.3.3
nav2-msgs==1.3.3
nav2-simple-commander==1.0.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
object-recognition-msgs==2.0.0
octomap-msgs==2.0.0
opencv-python==4.11.0.86
osrf-pycommon==2.1.4
packaging==25.0
pandas==2.2.3
pcl-msgs==1.0.0
pendulum-msgs==0.33.4
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-qt-binding==2.2.1
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.7.4
qt-gui==2.7.4
qt-gui-cpp==2.7.4
qt-gui-py-common==2.7.4
quality-of-service-demo-py==0.33.4
rcl-interfaces==2.0.2
rclpy==7.1.1
rcutils==6.7.1
realsense2-camera-msgs==4.55.1
regex==2024.11.6
requests==2.32.3
resource-retriever==3.4.3
rmw-dds-common==3.1.0
robot-localization==3.8.1
ros-gz-bridge==1.0.5
ros-gz-interfaces==1.0.5
ros-gz-sim==1.0.5
ros2action==0.32.1
ros2bag==0.26.4
ros2bag-mcap-cli==0.26.4
ros2bag-sqlite3-cli==0.26.4
ros2cli==0.32.1
ros2component==0.32.1
ros2controlcli==4.15.0
ros2doctor==0.32.1
ros2interface==0.32.1
ros2launch==0.26.5
ros2lifecycle==0.32.1
ros2multicast==0.32.1
ros2node==0.32.1
ros2param==0.32.1
ros2pkg==0.32.1
ros2run==0.32.1
ros2service==0.32.1
ros2topic==0.32.1
rosbag2-interfaces==0.26.4
rosbag2-py==0.26.4
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.3
rosidl-cli==4.6.3
rosidl-cmake==4.6.3
rosidl-generator-c==4.6.3
rosidl-generator-cpp==4.6.3
rosidl-generator-py==0.22.0
rosidl-generator-type-description==4.6.3
rosidl-parser==4.6.3
rosidl-pycommon==4.6.3
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.0
rosidl-typesupport-fastrtps-cpp==3.6.0
rosidl-typesupport-introspection-c==4.6.3
rosidl-typesupport-introspection-cpp==4.6.3
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.3
rqt-bag-plugins==1.5.3
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.0
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-tf-tree==1.0.5
rqt-topic==1.7.2
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
sensor-msgs==5.3.5
sensor-msgs-py==5.3.5
service-msgs==2.0.2
setuptools==79.0.1
shape-msgs==5.3.5
simpleaudio==1.0.4
six==1.17.0
slam-toolbox==2.8.1
smclib==4.1.0
srdfdom==2.0.4
sros2==0.13.2
statistics-msgs==2.0.2
std-msgs==5.3.5
std-srvs==5.3.5
stereo-msgs==5.3.5
sympy==1.13.3
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.36.4
tf2-kdl==0.36.4
tf2-msgs==0.36.4
tf2-py==0.36.4
tf2-ros-py==0.36.4
tf2-sensor-msgs==0.36.4
tf2-tools==0.36.4
theora-image-transport==4.0.3
tokenizers==0.21.1
topic-monitor==0.33.4
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
trajectory-msgs==5.3.5
transformers==4.51.3
triton==3.3.0
turtlebot3-msgs==2.2.1
turtlesim==1.8.3
type-description-interfaces==2.0.2
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.119
ultralytics-thop==2.0.14
unique-identifier-msgs==2.5.0
urdfdom-py==1.2.1
urllib3==2.4.0
vision-msgs==4.1.1
visualization-msgs==5.3.5
xacro==2.0.11
zipp==3.21.0
